requirements.txt
fastapi==0.105.0
uvicorn==0.23.2
SQLAlchemy==2.0.23
python-dotenv==1.0.0
pydantic-settings==2.1.0
psycopg2==2.9.9
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
redis==5.0.1
pytest==7.4.3
httpx==0.26.0
coverage==7.3.4
